Dang forgot to change that one

+++ src/translation.cpp
@@ -42,7 +42,7 @@ void emuInput() {
 // Renders a sprite from the given sprite sheet
 // at the given coordinates on the actual screen
 template <size_t sprites, size_t rows, size_t columns>
-void emuSprtRender(int sprt, int x, int y, bool flipX, bool flipY, const int (&sheet)[sprites][rows][columns], int colorOverride) {
+void emuSprtRender(int sprt, int x, int y, bool flipX, bool flipY, const uint8_t (&sheet)[sprites][rows][columns], int colorOverride) {
     // Don't render empty sprites or sprites outside the sheet
 	static const int noSprite[8][8] = {0};
     if (!memcmp(sheet[sprt], noSprite, sizeof(noSprite)) ||
